Started loading into d11
Started the process of loading the info from the OBJ parser into the d3d chain of life. Realized I actually need a reference to the device. Started to implenent the necessary info for setting the executable up into a dll.

+++ OBJLoader/code/obj_loader.h
@@ -7,17 +7,17 @@ struct obj
     r32* vertices;
     i32 vertexCount;
     i32 vertexLastIndex;
-    i32* vertexIndices;
+    u16* vertexIndices;
     
     r32* vertexNormals;
     i32 vertexNormalCount;
     i32 vertexNormalLastIndex;
-    i32* vertexNormalIndices;
+    u16* vertexNormalIndices;
     
     r32* vertexTextureCoordinates;
     i32 vertexTextureCoordCount;
     i32 vertexTextureCoordLastIndex;
-    i32* vertexTextureCoordIndices;
+    u16* vertexTextureCoordIndices;
     
     i32* face;
     i32 faceCount;
